Refactored FirmwareUpgradeError into its own file
Not many lines of code, but it's an important type to have on its own.
